Added realpath lib and fixed some stuff in bootstrap

+++ bootstrap.sh
@@ -1,15 +1,31 @@
 #! /usr/bin/env bash
 
+SCRIPT_DIR=`dirname $BASH_SOURCE`
+REALPATH_DIR=$SCRIPT_DIR/lib/realpath/realpath-lib
+source $REALPATH_DIR
+
+FILES=('.bashrc' '.bash_profile' '.gitconfig' 'z.sh')
+
+
+
 function set_os() {
     if [[ $(uname -a) == *"inux"* ]]; then
-        is_linux=true
-        is_mac=false
+        IS_LINUX=true
+        IS_MAC=false
     elif [[ $(uname -a) == *"arwin"* ]]; then
-        is_mac=true
-        is_linux=false
+        IS_MAC=true
+        IS_LINUX=false
     fi
 }
 set_os
+
+function debug() {
+    echo -e "SCRIPT_DIR      $SCRIPT_DIR"
+    echo -e "REALPATH_DIR    $REALPATH_DIR"
+    echo -e "IS_LINUX        $IS_LINUX"
+    echo -e "IS_MAC          $IS_MAC"
+    echo -e "Files that will be linked:\n`for FILE in ${FILES[@]}; do echo -e "  $FILE"; done`"
+}
 
 # Set used colors in bash
 GREEN=$(tput setaf 10)
@@ -26,6 +42,8 @@ usage ${GREEN} `basename $0` ${RESET}-d [folder_with_dofiles]"
 Options:
 -d\tInstall dotfiles in this dir (acctually creates symlinks)
 -r\tRemoves the symlinks restoring your backups (if they exist)
+-z\tDebug the info passed to make sure is correct use with -d [folder]
+\tEx.: ${GREEN} `basename $0` ${RESET}-zd [folder_with_dofiles]
 -h\tThis help
 "
 
@@ -49,32 +67,40 @@ function install() {
         echo ${GREEN}"Backing up your .gitconfig"
         mv .gitconfig .gitconfig_backup
     fi
+    if [ -f z.sh ]; then
+        echo ${GREEN}"Backing up your z.sh"
+        mv z.sh z.sh_backup
+    fi
 
     echo ${YELLOW}"Creating symlinks${RESET}
 `pwd`/.bash_profile -> $1/.bash_profile
-`pwd`/.bashrc -> $1/.bashrc"
+`pwd`/.bashrc -> $1/.bashrc
+`pwd`/.gitconfig -> $1/.gitconfig
+`pwd`/z.sh -> $1/z.sh"
     ln -s "$1/.bash_profile" .bash_profile
     ln -s "$1/.bashrc" .bashrc
     ln -s "$1/.gitconfig" .gitconfig
+    ln -s "$1/z.sh" z.sh
 
     echo "${GREEN}Installation finished.${RESET}
 run 'source .bash_profile' or open another terminal to see the changes"
 }
 
 function remove() {
-    if [ -h .bash_profile ] && [ -h .bashrc ]; then
+    if [ -h .bash_profile ] && [ -h .bashrc ] && [ -h .gitconfig ] && [ -h z.sh ]; then
         echo "${RED}Removing symlinks${RESET}"
-        rm -f .bash_profile .bashrc .gitconfig
-        if [ -f .bash_profile_backup ] && [ -f .bashrc_backup ] && [ -f .gitconfig_backup ]; then
+        rm -f .bash_profile .bashrc .gitconfig z.sh
+        if [ -f .bash_profile_backup ] && [ -f .bashrc_backup ] && [ -f .gitconfig_backup ] && [ -f z.sh ]; then
             echo "${GREEN}Restoring your backups"
             mv .bash_profile_backup .bash_profile
             mv .bashrc_backup .bashrc
             mv .gitconfig_backup .gitconfig
+            mv z.sh_backup z.sh
         else
             echo "Your backups where ${RED}NOT${RESET} restored"
         fi
     else
-        echo "The files .bash_profile and .bashrc ${RED}ARE NOT${RESET} symlinks, aborting"
+        echo "The files in home ${RED}ARE NOT${RESET} symlinks, aborting"
     fi
     echo "${GREEN}Removal completed.${RESET}
 run 'exec bash' or open another terminal to see the changes"
@@ -82,7 +108,7 @@ run 'exec bash' or open another terminal to see the changes"
 
 function rollback() {
     cd
-    if [ ! -f .bash_profile_backup ] && [ ! -f .bashrc_backup ] && [ ! -f .gitconfig_backup ]; then
+    if [ ! -f .bash_profile_backup ] && [ ! -f .bashrc_backup ] && [ ! -f .gitconfig_backup ] && [ ! -f z.sh ]; then
         echo "${RED}COULD NOT${RESET} find your backups"
         PS3="Restore anyway? "
         options=("Yes" "No")
@@ -99,17 +125,11 @@ function rollback() {
     fi
 }
 
-if $is_mac; then
-    function realpath() {
-        [[ $1 = /* ]] && echo "$1" || echo "$PWD/${1#./}" | sed -e "s/\/\/$//" -e "s/\/$//"
-    }
-fi
-
-if $is_linux; then
-    TEMP=`getopt -o hrd: --long help,remove,dir: -- "$@"`
-elif $is_mac; then
-    getopts hrd: TEMP "$@"
-fi
+#if $IS_LINUX; then
+#    TEMP=`getopt -o hrzd: --long help,remove,dir: -- "$@"`
+#elif $IS_MAC; then
+    getopts hrzd: TEMP "$@"
+#fi
 
 if [ $? != 0 ] || [ $# == 0 ]; then usage; exit 2; fi
 
@@ -121,17 +141,14 @@ do
     in
         (-d|d)
             if [ -d $2 ]; then
-                if $is_mac; then
-                    path=`realpath $OPTARG`
-                else
-                    path=`realpath $2`
-                fi
+                path=`get_dirname $OPTARG`
                 install $path
                 shift; shift
             fi
             ;;
         (-h|h) usage; exit 1 ;;
         (-r|r) rollback; exit 1 ;;
+        (-z|z) debug; exit 1 ;;
         (--) echo "shift"; shift; break ;;
     esac
 done
